Fixes problem with the "pick" button on show all template
When you picked a pokemon or item without previously inspecting it,
the base card was not being removed and inspecting another card would
remove the previously selected pokemon or item.

I added data-target and second-data-target to the pick button and
altered the function as is done for the add function in the inspect
modal. All seems to be working now.